Add suffix to store files
This prevents file/directory clashes. Example:
when we have a `tags` item, and a `tags/foo` item,
there will be a clash since the store creates:

- a file `store/tags`;
- a file `store/tags/foo`.

The second file requires the first file to be a
directory.

We simply solve this by adding a suffix to all
store files, so it becomes:

- a file `store/tags.hakyllstore`;
- a file `store/tags/foo.hakyllstore`.