shave sub-second precision off modification time
It seems like System.Directory's `getModificationTime` is now providing
sub-second precision if it's available in the underlying system call,
which it seems to be for most systems.

The problem is that when the modification time is stored, the DiffTime,
the number of seconds from midnight, has its sub-second precision
shaved off, effectively becoming 0 the next time it's read.  When a
file is then checked to see if it has been modified, the file is
_always_ deemed to have been modified because this time still includes
the sub-second precision, which will almost always be later in time
than 0, unless of course it happened to be 0 to begin with.

This simply shaves the sub-second precision off of the time returned by
`getModificationTime`, and thus no longer rebuilds every single file
each time any one file is saved.

Closes #250