Metadata: Allow org-style metadata headers
Org-style metadata definitions take the form of `#+KEY: value` and are
the standard way of defining metadata in Emacs Org-mode documents.  Such
as to avoid duplicate metadata definitions and a syntax unfamiliar to
org-mode users, this patch allowes the use of org-style metadata
definitions.  Every line at the beginning of a file which follows the
above pattern is treated as Hakyll metadata; keys are transformed to
lowercase.  The metadata-block is terminated by any other line.

As a consequence, if a user wants to define additional
metadata (e.g. `#+LINK` headers for Pandoc), she must separate those
metadata definitions from the topmost metadata block, e.g. using a blank
line.