Fix integer fields in YAML metadata
If a numeric field contains an integer number, it should be shown
as an integer number.  Without this patch, the field:

    answer = 42

would be shown as "42.0".  Before the introduction of YAML in
metadata, this was treated as a string and therefore shown as
just "42".  This patch reinstates the older (and IMHO correct)
behavior.