Expose settings for the preview server
This enables customization of the settings used by the preview server,
for example to implement URL rewriting.

My particular use case is to mimic the `tryFiles` setting in nginx to
add an extension when looking for a matching file. Then the URL can be
`/projects/web` and it will match a file called `projects/web.html` --
a nice clean URL that doesn't require lots of subdirectories and every
file being called `index.html`.

(This change doesn't implement that feature directly as part of
Hakyll, it just exposes the settings to enable building features like
it in your `site.hs`.)