runtime: simplify Runtime type (RWST → ReaderT) (#870)
As part of the implementation of the concurrent runtime, the state
parameter of the `Runtime` `RWST` changed from `RuntimeState` to
`MVar RuntimeState`.  As a result, that value is now read-only.
Furthermore, the writer parameter is unused (`()`).

Therefore, we do not need `RWST`; `ReaderT` suffices.  Add a new
field to `RuntimeRead` to hold the `MVar RuntimeState`, and populate
it.  Change `Runtime` to `ReaderT` and update the `getRuntimeState`
and `modifyRuntimeState` functions to reference the new field.