tolerate unexpected cache miss
If a Hakyll program termiantes abnormally (e.g. due to kill signal),
a subsequent run can fail because expected cache entries are not
present.  The missing cache entry is regarded as corruption; Hakyll
advises the user to delete the cache and start over.

But this scenario is recoverable, and the user need not be bothered.
If `Store.get` returns `NotFound`, don't fail - compile the item!
Other results (e.g. `WrongType`) still throw an error.