Hakyll.Web.Html: fix processing of `srcset` attribute
We used to assume that `srcset` is just like `src`, and contains only
one URL. As the name of the attribute might suggest, this assumption is
wrong.

The attribute's value has a somewhat-complex grammar, with optional
fields and different separators, so I implemented a Parsec parser for
(the most of) it. The parser expects URLs to be a contiguous string of
anything but space or comma (`srcset`'s separators); I hope it will be
sufficient.

This also adds tests for the affected functions, and documentation for
`getUrls`.

Fixes #889.