Add variants of pandoc renderer and compiler for items (#1020)
Add renderPandocItemWithTransformM and pandocItemCompilerWithTransformM,
which work like the respective functions without the `item' infix, but
the transformation function is an Item Pandoc -> Item Pandoc instead of
just a Pandoc -> Pandoc. This allows one to more seamlessly compose
functions that do require the extra information that an item provides,
like bibliographic transformations.