Validate XML of generated Atom/RSS feeds
Contents of the metadata fields of posts are included verbatim in the
generated Atom/RSS xml. This is intentional (we want to allow setting
markup tags in e.g. a post's title), but previously made it easy to
accidentally have hakyll produce invalid XML in feeds if any field
contained invalid XML.

Now instead these will be thrown as a compiler error.