Add example website (#1082)
* Add example website

I would like to propose adding my website to the list of Hakyll usage examples.  
It demonstrates that Hakyll can be used to build a structure where multiple blogs coexist on a single site.  
I believe this could serve as a useful example for other users, so I would appreciate it if you could consider including it.

* Update examples.markdown