Refactor the Dependency type to allow metadata-only dependencies
This is an API-breaking change, it refactors the old Dependency type
into three types: A Dependency, a DependencyKind, and a
DependencySelector. Through the DependencyKind we can specify a
dependency on either the Metadata or the Content of another Identifier.

The DependencySelector allows us, for both DependencyKinds, to either
depend on a single Identifier or a Pattern (as before). Dependencies of
KindMetadata are only considered out-of-date if the metadata changed.

This is extremely useful in many scenarios where Hakyll previously
induced unnecessary rebuilds because it could not distinguish between a
content and a metadata dependency. As an example, consider a navigation
sidebar displaying all available tags. There are also examples in the
existing code base where I have now used a metadata-only dependency
(getRoute, compilerGetMetadata, buildTagsWith).