Provider: Classify identifiers for which only the metadata changed
This is the foundation work required to introduce dependencies only on
metadata. Often times, a post's body/content may change, but its
metadata remains unmodified. Some compilers only need to be re-run if
the metadata changes (e.g., consider a sidebar that includes all post
tags). Presently, Hakyll does not allow depending on metadata, thereby
incurring a lot of unnecessary in such a scenario.

This commit allows classification of identifiers for which only the
metadata changed. The next commit will add an API to only depend on the
metadata of an identifier.

In technical terms it is implemented as follows: When invalidating the
metadataCache, retrieve the old metadata values. Afterward, for
identifiers modified (according to resourceModified), load the new
metadata and compare it to the old. Only consider metadata to have
changed if it differs.