Add support for declaring metadata-only dependencies (#1084)
* Provider: Classify identifiers for which only the metadata changed

This is the foundation work required to introduce dependencies only on
metadata. Often times, a post's body/content may change, but its
metadata remains unmodified. Some compilers only need to be re-run if
the metadata changes (e.g., consider a sidebar that includes all post
tags). Presently, Hakyll does not allow depending on metadata, thereby
incurring a lot of unnecessary in such a scenario.

This commit allows classification of identifiers for which only the
metadata changed. The next commit will add an API to only depend on the
metadata of an identifier.

In technical terms it is implemented as follows: When invalidating the
metadataCache, retrieve the old metadata values. Afterward, for
identifiers modified (according to resourceModified), load the new
metadata and compare it to the old. Only consider metadata to have
changed if it differs.

* Refactor the Dependency type to allow metadata-only dependencies

This is an API-breaking change, it refactors the old Dependency type
into three types: A Dependency, a DependencyKind, and a
DependencySelector. Through the DependencyKind we can specify a
dependency on either the Metadata or the Content of another Identifier.

The DependencySelector allows us, for both DependencyKinds, to either
depend on a single Identifier or a Pattern (as before). Dependencies of
KindMetadata are only considered out-of-date if the metadata changed.

This is extremely useful in many scenarios where Hakyll previously
induced unnecessary rebuilds because it could not distinguish between a
content and a metadata dependency. As an example, consider a navigation
sidebar displaying all available tags. There are also examples in the
existing code base where I have now used a metadata-only dependency
(getRoute, compilerGetMetadata, buildTagsWith).