getAllMetadata: Depend only on Metadata, not on Content (#1085)
This refines the dependency tracking of the getAllMetadata function
using e4b2dfd7d7fe8a6be163638efc8f7f0c9558afbb to only depend on the
metadata of matched Items.

For this purpose, it overwrites the default implementation of the
getAllMetadata function of the MonadMetadata type class in the
Compiler Monad instance.